+++ test/Unit/MabExperimentation/AutomatedExperimentTest.ts
@@ -23,30 +23,30 @@ describe('AutomatedExperimentTest', () => {
         baz: FooExperimentDeclaration.baz.ACTION1
     };
 
-    const ButtonAnimationsExperiment = new AutomatedExperiment({
+    const EndScreenExperiment = new AutomatedExperiment({
         actions: FooExperimentDeclaration,
         defaultActions: FooExperimentDefaultActions,
         cacheDisabled: true
     });
 
     it('should validate experiment action choices', () => {
-        assert.isTrue(ButtonAnimationsExperiment.isValid({ bar: 'bar1', baz: 'baz1' }));
-        assert.isTrue(ButtonAnimationsExperiment.isValid({ bar: 'bar1', baz: 'baz2' }));
-        assert.isTrue(ButtonAnimationsExperiment.isValid({ bar: 'bar2', baz: 'baz1' }));
-        assert.isTrue(ButtonAnimationsExperiment.isValid({ bar: 'bar2', baz: 'baz2' }));
-        assert.isTrue(ButtonAnimationsExperiment.isValid({ bar: 'bar2', baz: 'baz2', optional: 'defined' }));
+        assert.isTrue(EndScreenExperiment.isValid({ bar: 'bar1', baz: 'baz1' }));
+        assert.isTrue(EndScreenExperiment.isValid({ bar: 'bar1', baz: 'baz2' }));
+        assert.isTrue(EndScreenExperiment.isValid({ bar: 'bar2', baz: 'baz1' }));
+        assert.isTrue(EndScreenExperiment.isValid({ bar: 'bar2', baz: 'baz2' }));
+        assert.isTrue(EndScreenExperiment.isValid({ bar: 'bar2', baz: 'baz2', optional: 'defined' }));
     });
 
     it('should fail validation on invalid experiment action choices', () => {
-        assert.isFalse(ButtonAnimationsExperiment.isValid({}));
-        assert.isFalse(ButtonAnimationsExperiment.isValid({ bar: 'bar1' }));
-        assert.isFalse(ButtonAnimationsExperiment.isValid({ baz: 'baz1' }));
-        assert.isFalse(ButtonAnimationsExperiment.isValid({ baz: 'invalid' }));
-        assert.isFalse(ButtonAnimationsExperiment.isValid({ someOtherField: 'sad' }));
-        assert.isFalse(ButtonAnimationsExperiment.isValid({ bar: 'bar1', baz: 'invalid', someOtherField: 'baz1' }));
+        assert.isFalse(EndScreenExperiment.isValid({}));
+        assert.isFalse(EndScreenExperiment.isValid({ bar: 'bar1' }));
+        assert.isFalse(EndScreenExperiment.isValid({ baz: 'baz1' }));
+        assert.isFalse(EndScreenExperiment.isValid({ baz: 'invalid' }));
+        assert.isFalse(EndScreenExperiment.isValid({ someOtherField: 'sad' }));
+        assert.isFalse(EndScreenExperiment.isValid({ bar: 'bar1', baz: 'invalid', someOtherField: 'baz1' }));
     });
 
     it('should validate experiment action choices with extra fields', () => {
-        assert.isTrue(ButtonAnimationsExperiment.isValid({ bar: 'bar1', baz: 'baz1', extra: 'sad' }));
+        assert.isTrue(EndScreenExperiment.isValid({ bar: 'bar1', baz: 'baz1', extra: 'sad' }));
     });
 });